+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2026-08-21
+
 ### Fixed
 
 - `Connect-ZTAssessment -UseDeviceCode` printed nothing to the console when